move dockerEventBundle class to its own file for readability
